move showError and showWarning in a .cpp file
