+++ plugins/builtin/include/content/helpers/notification.hpp
@@ -4,13 +4,7 @@
 
 namespace hex::plugin::builtin {
 
-    void showError(const std::string& message){
-        PopupError::open(message);
-        log::error(message);
-    }
+    void showError(const std::string& message);
 
-    void showWarning(const std::string& message){
-        PopupWarning::open(message);
-        log::warn(message);
-    }
+    void showWarning(const std::string& message);
 }
