+++ plugins/builtin/source/content/helpers/notification.cpp
@@ -1,0 +1,16 @@
+#include <hex/helpers/logger.hpp>
+
+#include <content/popups/popup_notification.hpp>
+
+namespace hex::plugin::builtin {
+
+    void showError(const std::string& message){
+        PopupError::open(message);
+        log::error(message);
+    }
+
+    void showWarning(const std::string& message){
+        PopupWarning::open(message);
+        log::warn(message);
+    }
+}
